Let note and lyric Selection blueprint inherit same class.
