+++ osu.Game.Rulesets.Karaoke/Edit/Blueprints/KaraokeSelectionBlueprint.cs
@@ -3,39 +3,22 @@
 
 using osu.Framework.Graphics;
 using osu.Game.Rulesets.Edit;
+using osu.Game.Rulesets.Karaoke.Objects;
 using osu.Game.Rulesets.Karaoke.Objects.Drawables;
 using osu.Game.Rulesets.Objects.Drawables;
 using osuTK;
 
 namespace osu.Game.Rulesets.Karaoke.Edit.Blueprints
 {
-    public class KaraokeSelectionBlueprint : OverlaySelectionBlueprint
+    public class KaraokeSelectionBlueprint<T> : OverlaySelectionBlueprint
+        where T : KaraokeHitObject
     {
-        public new DrawableKaraokeScrollingHitObject DrawableObject => (DrawableKaraokeScrollingHitObject)base.DrawableObject;
+        protected new T HitObject => (T)DrawableObject.HitObject;
 
         public KaraokeSelectionBlueprint(DrawableHitObject drawableObject)
             : base(drawableObject)
         {
             RelativeSizeAxes = Axes.None;
-        }
-
-        protected override void Update()
-        {
-            base.Update();
-
-            Position = Parent.ToLocalSpace(DrawableObject.ToScreenSpace(Vector2.Zero));
-        }
-
-        public override void Show()
-        {
-            DrawableObject.AlwaysAlive = true;
-            base.Show();
-        }
-
-        public override void Hide()
-        {
-            DrawableObject.AlwaysAlive = false;
-            base.Hide();
         }
     }
 }
